+++ tests/test_match.py
@@ -1,7 +1,7 @@
 # type: ignore
-from vlrscraper.match import Match, PlayerStats
-from vlrscraper.team import Team
 from vlrscraper.utils import previous_epoch
+from vlrscraper.controllers import MatchController
+from vlrscraper.resources import Team, PlayerStats, Match
 
 from .helpers import assert_teams
 
@@ -48,12 +48,12 @@ def test_match_eq():
 
 
 def test_match_player_get_ids(requests_regression):
-    m = Match.get_player_match_ids(4004, previous_epoch(days=30))
+    m = MatchController.get_player_match_ids(4004, previous_epoch(days=30))
     assert m == [413228, 413189, 412065, 408415, 408414]
 
     assert (
         len(
-            Match.get_player_match_ids(
+            MatchController.get_player_match_ids(
                 4004, previous_epoch(days=60), previous_epoch(days=30)
             )
         )
@@ -62,7 +62,7 @@ def test_match_player_get_ids(requests_regression):
 
     assert (
         len(
-            Match.get_player_match_ids(
+            MatchController.get_player_match_ids(
                 4004, previous_epoch(days=30), previous_epoch(days=30)
             )
         )
@@ -71,12 +71,12 @@ def test_match_player_get_ids(requests_regression):
 
 
 def test_match_team_get_ids(requests_regression):
-    m = Match.get_team_match_ids(2, previous_epoch(days=30))
+    m = MatchController.get_team_match_ids(2, previous_epoch(days=30))
     assert m == [412065, 408415, 408414]
 
     assert (
         len(
-            Match.get_team_match_ids(
+            MatchController.get_team_match_ids(
                 2, previous_epoch(days=60), previous_epoch(days=30)
             )
         )
@@ -84,7 +84,7 @@ def test_match_team_get_ids(requests_regression):
     )
     assert (
         len(
-            Match.get_team_match_ids(
+            MatchController.get_team_match_ids(
                 2, previous_epoch(days=30), previous_epoch(days=30)
             )
         )
@@ -93,7 +93,7 @@ def test_match_team_get_ids(requests_regression):
 
 
 def test_match_player_get(requests_regression):
-    matches = Match.get_player_matches(4004, previous_epoch(days=30))
+    matches = MatchController.get_player_matches(4004, previous_epoch(days=30))
     assert len(matches) == 5
     assert matches[3].get_player_stats(729) == PlayerStats(
         1.2, 245, 39, 30, 21, 9, 78, 146, 25, 9, 4, 5
@@ -105,7 +105,7 @@ def test_match_player_get(requests_regression):
 
 def test_match_get():
     # Current match
-    m = Match.get_match(408415)
+    m = MatchController.get_match(408415)
     assert m is not None
     assert m.get_id() == 408415
     assert m.get_name() == "NA Play-ins: Grand Final"
@@ -127,18 +127,18 @@ def test_match_get():
     )
 
     # China match (no stats)
-    m = Match.get_match(370727)
+    m = MatchController.get_match(370727)
     assert m is not None
     assert m.get_player_stats(3520) == PlayerStats(
         None, 204, 77, 82, 20, -5, None, 131, 24, 18, 25, -7
     )
 
     # Old match (no stats)
-    m = Match.get_match(3490)
+    m = MatchController.get_match(3490)
     assert m is not None
     assert m.get_player_stats(4004) == PlayerStats(
         None, 176, 13, 17, 8, -4, None, 112, 28, 2, 1, 1
     )
 
-    assert Match.get_match(0) is None
-    assert Match.get_match("3490") is None
+    assert MatchController.get_match(0) is None
+    assert MatchController.get_match("3490") is None
